shuffle output back to weave
